% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allBAF_FDR.R
\name{allBAF_FDR}
\alias{allBAF_FDR}
\title{Function to calculate the FDR and p-value from the BAF data}
\usage{
allBAF_FDR(data, epsilon_cut = 0.52, ...)
}
\arguments{
\item{data}{Data must be a GRanges object with several meta-data columns.  The GRange itself gives the position of each SNP.  Two columns, Group and Region (which typically refer to a "Sample" and "Gene" respectively) give which region and group the SNP belongs to.  CorrectedPhase should give (as a character) the phase of each SNP and phaseOfMajorAllele should give which phase at that SNP belongs to the chromosome with the highest CN.  MUT_ALT should given the number of reads supporting the mutant allele for the SNP and MUT should give the total number of reads covering the SNP.  Finally, tau, should give the bias correction factor for each SNP.}

\item{epsilon_cut}{The null hypothesis is that 0.5<epsilon<epsilon_cut.  Determine this by looking at the fitted values for epsilon and adjusting so we exclude the exponential at epsilon=0.5.  0.52 seems to work pretty well in most cases.}

\item{...}{Passed to baf_phased_grid_ml.}
}
\value{
A table giving the p-values, ML estimates of epsilon and FDR for each pair of "Group" and "Region" for which we have data.
}
\description{
Given a GRanges object containing bi-allelic frequency data for heterozygous SNPs, calculates the false discovery rate and p-value for rejecting the null hypothesis that the alleles are in balance at each Region/Group (Gene/Patient) combination.
}

