% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allCoverageFDR.R
\name{allCoverageFDR}
\alias{allCoverageFDR}
\title{Calculates the coverage FDR for all patients}
\usage{
allCoverageFDR(patMap, ...)
}
\arguments{
\item{patMap}{A data.frame with columns Patient, Sample, and Type. Type should be 'Tumour' and 'Normal'.  Patient is the group designator and Sample identifies individual samples.}

\item{...}{Extra arguments passed to coverageFDR}
}
\description{
Given a data frame describing how samples are grouped together, calculates p-values and false discovery rates for the rejection of the null hypothesis that the copy number state in each "Region"/"Group" (Gene/Patient) pair is equal to the ploidy.
}

