% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFDR.R
\name{allFDR}
\alias{allFDR}
\title{Performs combined inference on coverage and BAF data.}
\usage{
allFDR(bafp, patMap, ...)
}
\arguments{
\item{bafp}{Output from allBAF_FDR}

\item{patMap}{See help for allCoverageFDR}

\item{...}{Extra arguments passed to coverageFDR}
}
\value{
The bafp data.frame with extra columns combLR, giving the combined likelihood ratio for the patient/gene combination in a given row, and comb_pval, giving the p-value for the same.
}
\description{
Given the BAF based calls for difference from allelic balance, combines these with likelihood ratios for the coverage data for the same samples/genes to obtain a likelihood ratio that compares the null hypothesis of copy neutral allelic balance against the alternative of some difference.
}

