% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baf_phased_grid_ml.R
\name{baf_phased_grid_ml}
\alias{baf_phased_grid_ml}
\title{Simultaneously calculates the maximum likelihood estimate of epsilon, the ratio of major allele CN to total CN for all groups and regions.}
\usage{
baf_phased_grid_ml(data, mask, left = 0, right = 1, tol = 1e-06,
  ngrid = 12, verbose = TRUE)
}
\arguments{
\item{data}{The BAF data table.}

\item{mask}{A mask which tells us the rows of the data table to group together.  Usually interaction(data$Group,data$Region,drop=TRUE) or something similar.}

\item{left}{The starting lower-bound estimate for epsilon.  Can be the same length as data, but if of length 1 all data points get the same starting lower-bound.}

\item{right}{The starting upper-bound estimate for epsilon.  Can be the same length as data, but if of length 1 all data points get the same starting upper-bound.}

\item{tol}{Search will terminate when all values of epsilon are known to at least this accuracy.}

\item{ngrid}{Split the space between left and right into this many pieces each iteration.}

\item{verbose}{Print progress and diagnostic messages.}
}
\value{
A list containing par, the parameter estimates and value the negative log-likelihood values of those estimates.
}
\description{
This is obviously less efficient than a version based on optim on a case by case basis, but with this approach the whole problem can be vectorised, making it massively faster.  Assumes a convex shape to the likelihood function with a maximum somewhere in the range 0 < epsilon < 1.
}

