% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCorrectionFactors.R
\name{calcCorrectionFactors}
\alias{calcCorrectionFactors}
\title{Calculates a correction factor for each sample.}
\usage{
calcCorrectionFactors(counts, chi, null_rows, trim = NA)
}
\arguments{
\item{counts}{A matrix containing the number of fragments in each group/region, with columns giving samples and rows giving region (gene).}

\item{chi}{A matrix with the same shape as counts but containing bias correction factors for each combination (usually counts predicted from GC content).}

\item{null_rows}{The rows in counts/chi that we will assume have CN = ploidy on average.}

\item{trim}{The amount of data to trim when calculating the trimmed mean.  If set to NA, use median instead.}
}
\value{
A vector named by samples giving the calculated correction factors.
}
\description{
Calculates a factor which can be used to normalise each sample to its ploidy, under the assumption that most of the regions specified in "null_rows" have copy number equal to the sample ploidy.
}

