% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBAF_GC_Correction.R
\name{calculateBAF_GC_Correction}
\alias{calculateBAF_GC_Correction}
\title{Calculates GC bias correction factors for each SNP}
\usage{
calculateBAF_GC_Correction(bafs, gcBias)
}
\arguments{
\item{bafs}{GRanges object containing data on the SNPs for the sample of interest.}

\item{gcBias}{The output of GCnorm for the sample of interest.}
}
\value{
The bafs object with an extra GC_correction column.
}
\description{
Uses GC bias curve to calculate the correction that should be applied in order to correct for the preference for the alternate/reference allele when the GC content changes between them.  That is, if the reference and alternate allele of a SNP have different GC content (e.g., G -> T), calculates a per-SNP correction factor to account for the change in predicted coverage for the alternate allele owing to the different GC content of the two alleles.
}

