% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCoverage.R
\name{calculateCoverage}
\alias{calculateCoverage}
\title{Get's the coverage at the locations specified.}
\usage{
calculateCoverage(bfile, tgts)
}
\arguments{
\item{bfile}{The BAM file to load data from.}

\item{tgts}{The GRanges at which to calculate the coverage.}
}
\value{
The tgts object with a obsCov column containing the coverage at each location.
}
\description{
Calculates the number of fragments overlapping each base specified from the paired-end data in the BAM file referenced. Should only be used for small regions, not large chunks of the genome.
}

