% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReads.R
\name{countReads}
\alias{countReads}
\title{Count the number of fragment 5' locations falling with specified target regions.}
\usage{
countReads(bfile, tgts)
}
\arguments{
\item{bfile}{The location of the BAM file to load data from.}

\item{tgts}{A GRanges object containing the ranges that we want to fetch counts for.}
}
\value{
The tgts GRange with extra columns forCounts, revCounts and unknownCounts giving the number of fragments with 5' in the target region by fragment orientation.
}
\description{
This is very conservative with the reads it includes, throwing out anything that doesn't look very solid.
}

