% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageFDR.R
\name{coverageFDR}
\alias{coverageFDR}
\title{Function to calculate the FDR and p-value from the coverage}
\usage{
coverageFDR(counts, chi, sample_correction, tgt_cols, norm_cols,
  lfc_cut = 0.15, deviance = TRUE)
}
\arguments{
\item{counts}{Table of counts containing the normal and tumour samples to compare.}

\item{chi}{Table of bias factors containing the same rows (in the same order) as counts and containing the normal and tumour samples to compare.}

\item{sample_correction}{Named vector, where names are samples and values are the sample correction for composition effects.  Usually the values returned by calcCorrectionFactors.}

\item{tgt_cols}{The column names or numbers for the target samples.}

\item{norm_cols}{The column names or numbers for the normal samples.}

\item{lfc_cut}{The log fold-change that you consider to be equivalent to no change.  Calibrate this by looking at the logFC values coming out of the fit and excluding those likely to be normal.  To be really sure, pick out those genes with epsilon < epsilon_cut and plot their logFC (most will be CN neutral).  0.15 seems reasonable in most cases.}

\item{deviance}{If True, returns the deviance for the fitted model (i.e., the difference in log likelihood between the fitted and null model.}
}
\value{
A edgeR glm fit object for each row having CN different from ploidy.
}
\description{
Calculates a false discovery rate for rejecting the null hypothesis that the copy number state is equal to the ploidy in each region.
}

