% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Finds the maximum likelihood values of purity and ploidy and the CN states for each gene.}
\usage{
fitModel(rho = 0.6, pl = 2, ...)
}
\arguments{
\item{rho}{Starting estimate for sample purity.}

\item{pl}{Starting estimate for sample ploidy.}

\item{...}{Parameters to be passed to nllPatient. You must not specify purities and ploidies.}
}
\value{
A list containing the fit object and the grid of nlls for different CN states for genes at the optim values of purity and ploidy.
}
\description{
This function attempts to determine the precise copy number state, ploidy and sample purity of each group and region (patient and gene).  This works by maximising the likelihood of a joint model combining both the coverage and BAF data.  Unless you have a large number of target regions (e.g., exome data), this estimate is likely to be biased towards solutions with unrealistically complex copy number states and low purities.
}

