% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_chr}
\alias{fix_chr}
\title{Standardise chromosome names}
\usage{
fix_chr(chrs)
}
\arguments{
\item{chrs}{The chromosomes to fix up.}
}
\value{
The cleaned up chromosome names.
}
\description{
Tuned for human, standardise chromosome names.
}

