% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCnorm.R
\name{GCnorm}
\alias{GCnorm}
\title{Estimates GC bias using the method of Benjamini and Speed (NAR 2011)}
\usage{
GCnorm(bfile, tgts, genome, lstrip = 3, rstrip = 3, gc_smooth = 0.03,
  maxReads = 1e+07)
}
\arguments{
\item{bfile}{The path to a BAM file containing the mapped reads.}

\item{tgts}{A GRanges object giving the locations of the genome to consider.}

\item{genome}{A BSgenome object containing the genome to consider.}

\item{lstrip}{The number of bases to trim from the 5' end of window in which GC is considered.}

\item{rstrip}{The number of bases to trim from the 3' end of window in which GC is considered.}

\item{gc_smooth}{Smooth the predicted rates in bins of this width.  If less than zero interpreted as fraction of insert size.}

\item{maxReads}{Passed to BAM file reading functions.  The maximum number of reads to load at a time to prevent memory exhaustion.}
}
\value{
A list containing:
forw - A vector giving the predicted number of forward strand fragments starting at a location with the GC content in a forward window given by the vector name.
forw - A vector giving the predicted number of reverse strand fragments starting at a location with the GC content in a reverse window given by the vector name.
isize - The median insert size of the data.  The window where GC content is counted is set to this size.
lstrip - The value passed to this function.
rstrip - The value passed to this function.
readLen - The median read length of the data.
genome - The value passed to this function.
bfile - The value passed to this function.
gc_smooth - The number of GC count values over which the predicted fragment rates was smoothed.
}
\description{
Uses the method of Benjamini and Speed to calculate rates at which we expect reads to occur based on GC content.  Returns predicted rates at which reads to occur on the forward and reverse strand, based on their GC content in a window starting at that position of length the insert size (with lstrip and rstrip bases trimmed from the 5' and 3' end respectively).
}
\details{
As well as a BAM file to extract fragment counts from, this function also needs a GRanges object defining which regions of the genome to consider.  This should be set to those bases that have an acceptably high mappability for your data (and are in the target region if using targeted data).
}

