% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GC_counts_for_region.R
\name{get_GC_counts_for_region}
\alias{get_GC_counts_for_region}
\title{Get's the GC content in windows anchored at the small number of genomic locations given}
\usage{
get_GC_counts_for_region(regions, isize, lstrip = 3, rstrip = 3,
  summarise = TRUE, genome)
}
\arguments{
\item{regions}{The GRanges object giving the positions at which GC should be calculated.}

\item{isize}{The insert size to use at each location (including lstrip and rstrip).}

\item{lstrip}{The number of bases skipped at the 5' end of the fragment.}

\item{rstrip}{The number of bases skipped at the 3' end of the fragment.}

\item{summarise}{If True, The total count of each GC value within each region is returned.  Otherwise, the raw GC values at each location are.}

\item{genome}{The BSgenome object from which data is extracted.}
}
\value{
A list containing forward and reverse strand counts under as "forw" and "rev".  If summarise is False, each will contain a list of the same length as regions, named by the regions, containing the GC count based at each position in regions in turn.  If summarised, each entry is a matrix with columns giving the number of GCs and rows giving the different regions.
}
\description{
Unlike get_GC_counts, this function is optimised to get the GC content at a small number of genomic locations.  It should probably never be directly called, instead use get_GC_counts which will call this function when appropriate.
}

