% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinNll.R
\name{getMinNll}
\alias{getMinNll}
\title{Calculates the minimum negative log-likelihood for each purity/ploidy value.}
\usage{
getMinNll(cngrid, calcWith = c("nllBAF", "nllCov"))
}
\arguments{
\item{cngrid}{The data.frame returned by nllPatient.}

\item{calcWith}{A vector giving the different contributions to the total negative log-likelihood to sum, minimise, and plot.}
}
\value{
A data.frame giving the minimum nll for each combination of purity and ploidy.
}
\description{
Given a table with all the possible CN states, returns the values which minimise the negative log-likelihood.
}

