% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormalisedCoverage.R
\name{getNormalisedCoverage}
\alias{getNormalisedCoverage}
\title{Calculates the normalised coverage in bins for a small set of target regions.}
\usage{
getNormalisedCoverage(bfile, ltgts, gcBias, sample_correction = 1,
  position_correction = 1, bin_size = 100, biasCorrect = "GPS")
}
\arguments{
\item{bfile}{The location of the BAM file to get the reads from.}

\item{ltgts}{A GRanges object specifying the locations to get data for.  Should be small compared to the size of the genome.}

\item{gcBias}{A gcBias object.}

\item{sample_correction}{The sample correction factor to be applied to the data.}

\item{position_correction}{The position corrections to apply to the data.}

\item{bin_size}{The size of the bins to average coverage in.}

\item{biasCorrect}{What type of bias correction to apply.  'G' in string = GC, 'S' in string = sample, 'P' in string = position.}
}
\value{
A data.frame giving the normalised coverage in bins across the target region.  An altIsMajor column is set, which oscillates every time there is a change between a GRanges row.
}
\description{
Here "coverage" means "number of fragments overlapping each position" when bin_size = NA and "number of 5' fragments falling in each bin" when bin_size is a positive integer.  In both cases normalisation is done using the bias factors specified in biasCorrect.
}

