% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCN_State.R
\name{inferCN_State}
\alias{inferCN_State}
\title{Combine and predict CN state}
\usage{
inferCN_State(covp, bafp, FDR_Cut = 0.01, SNP_Cutoff = 10,
  coverageNormCut = 0.06, coverageExtremeCut = 0.3)
}
\arguments{
\item{covp}{The output of coverageFDR.}

\item{bafp}{The output of allBAF_FDR.}

\item{FDR_Cut}{The FDR cut-off used in calling CN states.}

\item{SNP_Cutoff}{How many SNPs should a region have before we can confidently use the BAF data?}

\item{coverageNormCut}{Regions that have absolute log coverage (tumour/normal) less than this cutoff are considered to be CN neutral.}

\item{coverageExtremeCut}{Regions that have absolute log coverage (tumour/normal) greater than this cutoff are considered to have extreme CN.  This is only used to determine regions that are homozygously deleted or duplicated if we have significant BAF data in support of this conclusion already.}
}
\value{
A data.frame containing the calls for each type of data and interpretations of these calls.
}
\description{
Creates a series of human readable interpretations of the calls from the BAF and coverage based statistical models.  The level of detail provided in the interpretation differs depending on the number of SNPs present in each region, with more SNPs allowing for more detailed interpretation.  For example, when only a small number of SNPs are available, only "amplified", "deleted", and "neutral" are selected from.
}

