% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferHetSNPs.R
\name{inferHetSNPs}
\alias{inferHetSNPs}
\title{Detect SNPs heterozygous in normal.}
\usage{
inferHetSNPs(baf, seq_error_rate = 0.03, omega = 1, fdr_cut = 0.01,
  useNorm = TRUE)
}
\arguments{
\item{baf}{The GRanges BAF object.  Should already have columns REF,ALT,MUT,MUT_ALT,NORM,NORM_ALT,Group, and Region  NORM and NORM_ALT should be NA where there is no matched normal.}

\item{seq_error_rate}{The assumed rate at which bases appear in error when homozygous for a particular allele.}

\item{omega}{The degree of reference strand bias.}

\item{fdr_cut}{The FDR cut-off to use to declare a SNP is not homozygous.}

\item{useNorm}{Should we use the counts in the normal or the mutant?  A vector of length baf or a single entry.}
}
\value{
A logical vector of length = length(baf) indicating if a SNP is heterozygous in the normal.
}
\description{
Determine which SNPs from the input array are heterozygous based on the counts for reference and alternate allele in the tumour or in the normal if this information is given (i.e., if the sample has a matched normal).
This method works by constructing two p-values, one for getting the number of alternate reads if the SNP is homozygous for the reference and one for getting the number of reference reads if the SNP is homozygous for the alternate allele.  In an ideal world, the probability in both these cases would be 0 for any number of counts greater than 0.  However, errors cause this to be untrue.  So the p-value is calculated assuming a prescribed error rate.  Do not try and set this accurately, instead set it highly conservatively to exclude false positives.
}

