% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ltable}
\alias{ltable}
\title{An enhancement to the standard table function}
\usage{
ltable(..., levels = NULL, dropExtra = TRUE, defValue = 0)
}
\arguments{
\item{...}{The thing to be counted, passed to \code{table}.}

\item{levels}{These are entries that are always in the table output}

\item{dropExtra}{Should we drop any value not specified?}

\item{defValue}{Default value}
}
\value{
A named vector, the same as table but with all levels present and set to defValue if no data is present.
}
\description{
Pads out standard calls to table with extra levels not present in the data without the need to force input data to be factors.
}

