% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSunrise.R
\name{makeSunrise}
\alias{makeSunrise}
\title{Creates a sunrise plot for the sample specified.}
\usage{
makeSunrise(purities = seq(0, 1, length.out = 101), ploidies = seq(1, 12,
  length.out = 101), ...)
}
\arguments{
\item{purities}{The purities to test.}

\item{ploidies}{The ploidies to test.}

\item{...}{Extra parameters that are passed to nllPatient which is used to calculate the nlls.}
}
\value{
A ggplot object containing the sunrise plot.
}
\description{
Creates a sunrise plot comparing the total negative log-likelihood of models with different purities and ploidies.
}

