% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllBAF.R
\name{nllBAF}
\alias{nllBAF}
\title{Calculates the negative log-likelihood of the BAFs having the CN and purity values given in sgrid at the gene level.}
\usage{
nllBAF(bafs, sgrid)
}
\arguments{
\item{bafs}{The usual BAF GRanges object, the data from which is used to calculate likelihoods.}

\item{sgrid}{A data.frame with columns gene,nA,nB and, purity giving the combinations of gene, CN of major allele, CN of minor allele and, tumour purity to test.}
}
\value{
sgrid with an extra column called nllBAF containing the sum of the negative log-likeihoods for the combination of parameters listed in the row summed across all SNPs in bafs for that gene.
}
\description{
For each combination of gene,A,A+B,rho in sgrid, calculate the negative log-likelihood of a binomial model for the BAFs observed with these parameters.  This is nicely vectorised so, providing the memory doesn't become an issue, we can calculate all values of rho and A,A+B simultaneously.  The genes, purities and, CN states that are tested are only those that appear in sgrid.
We assume that all entries in bafs are from the same patient and so should be aggregated.  That is, if bafs contains SNP data from multiple samples, they will all be treated as if they came from one sample.  This is appropriate for samples which are technical or biological replicates, but obviously not for different conditions/patients.
}

