% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllCov.R
\name{nllCov}
\alias{nllCov}
\title{Calculate the negative log-likelihood of different CN/purity/ploidy combinations for the coverage data.}
\usage{
nllCov(counts, chi, sample_correction, tumours, normals, sgrid, disp = 0)
}
\arguments{
\item{counts}{The table of counts containing the genes and samples of interest.}

\item{chi}{The table of bias correction factor containing the genes and samples of interest.}

\item{sample_correction}{The vector containing the sample correction factors to apply.}

\item{tumours}{The names (or column numbers in counts/chi/sample_correction) of the tumour samples to test.}

\item{normals}{The names (or column numbers in counts/chi/sample_correction) of the normal samples to compare against.}

\item{sgrid}{A data.frame containing columns gene,nA,nB,purity and, ploidy.  These specify the combinations of factors to test and calculate the negative log-likelihood for.  nA is the CN of the major allele, nB is the CN of the minor allele.}

\item{disp}{An over-dispersion parameter (1/psi in the usual notation) that is added to the negative binomial model.}
}
\value{
sgrid with an extra column, nllCov, containing the negative log-likelihood estimates for each combination in sgrid.
}
\description{
This fits the glm over and over with different offsets, where those offsets are chosen to model the choice of different values of A (major copy number), B (minor copy number), purity and, ploidy.  To extract the negative log-likelihood from the model, the deviance from the model with one free parameter and with only an intercept term are compared.  What this means is that what we return is not actually the nll, but the nll minus some constant which is the nll of the null model.  For maximum likelihood estimation this should not matter, but be aware of this if you try and use it to construct a statistical test, or something else.
sgrid controls which CN states, purities and ploidies are tested.
}

