% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllPatient.R
\name{nllPatient}
\alias{nllPatient}
\title{Calculates the total negative log-likelihood from a model including both BAF and coverage data for a range of different CN states, purities and ploidies.}
\usage{
nllPatient(counts, chi, sample_correction, bafs, tumours, normals, disp = 0,
  n = 0:5, purities = seq(0, 1, length.out = 101), ploidies = seq(1, 6,
  length.out = 101), verbose = TRUE)
}
\arguments{
\item{counts}{The table of counts giving the number of reads with 5' fragments in regions of interest in different samples.}

\item{chi}{The table of bias correction factors for regions and samples of interest.}

\item{sample_correction}{The vector containing the sample correction factors to apply.}

\item{bafs}{The usual BAF GRanges object, the data from which is used to calculate likelihoods for the BAF model.}

\item{tumours}{The tumour samples to group together and calculate nll for.}

\item{normals}{The normal samples to compare against (only needed for coverage model).}

\item{disp}{A negative-binomial over-dispersion parameter used in the coverage model.  disp=1/psi in the usual notation.}

\item{n}{A range of CN states to test.  That is, nllPatient will calculate the nll for models where the CN state is set to all combinations of values given in n.}

\item{purities}{The range of purity values to calculate nll for.}

\item{ploidies}{The range of ploidy values to calculate nll for.}

\item{verbose}{Should we print progress and diagnostic messages?}
}
\value{
A data.frame giving the combinations of gene,CN state, purity and ploidy and the negative log-likelihoods for each combination.
}
\description{
For the samples specified, calculates the negative log-likelihood for the different combinations of CN values specified.
Essentially constructs a giant table, with each gene's CN state by allele, purity ploidy and nll for coverage and BAF models.  Will have size nrho*nploidy*(max_CN+1)*(max_CN+2)/2*nRegions.  In most cases this should not be called directly, but via fitModel.
}

