% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_chr}
\alias{num_chr}
\title{Make chromosome names into integers}
\usage{
num_chr(chrs)
}
\arguments{
\item{chrs}{The chromosomes to convert.}
}
\value{
The converted chromosomes as a numeric vector.
}
\description{
Converts chromosome names into integers, useful for sorting.
}

