% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseSNPs.R
\name{phaseSNPs}
\alias{phaseSNPs}
\title{Phases SNPs using IMPUTE.}
\usage{
phaseSNPs(bafs, impute_exec, impute_m, impute_h, impute_l, tgts = NULL,
  correctPhase = TRUE, tmpDir = ".", verbose = TRUE)
}
\arguments{
\item{bafs}{A GRanges object holding SNPs heterzygous in normal with columns as described above.}

\item{impute_exec}{The location of the IMPUTE executable.}

\item{impute_m}{A string giving the location of the impute m files where \%s is to be replaced by the chromosome name.}

\item{impute_h}{A string giving the location of the impute h files where \%s is to be replaced by the chromosome name.}

\item{impute_l}{A string giving the location of the impute l files where \%s is to be replaced by the chromosome name.}

\item{tgts}{A GRanges object with a column named regionCol which contains the GRanges spanned by the target region.  Used to provide limits to IMPUTE.  If NULL, these are set to the max/min location of SNPs in the data.}

\item{correctPhase}{Should we try and improve the phasing by toggling haplotype blocks that improve the variance in the estimate of both alleles CN.}

\item{tmpDir}{A directory we can write to where the temporary IMPUTE files will be created.}

\item{verbose}{Should we output diagnostic and progress messages.}
}
\value{
bafs object with extra columns holding the phasing information
}
\description{
The input is assumed to contain only SNPs heterozygous in the normal, for which we already have the core columns listed below.  IMPUTE is then run on each Region/Group combination to infer phasing.
}
\details{
The key columns that the BAF GRanges object needs to have are:
\itemize{ 
 \item REF - The reference base.
 \item ALT - The alternate base.
 \item MUT - The number of reads covering this location in the tumour.
 \item MUT_ALT - The number of reads with the alternate base at this location in the tumour.
 \item Group - String that will be used to aggregate SNPs across different sequencing experiments (samples usually).
 \item Region - Will be used to aggregate SNPs at different genomic locations (typically genes).
}
And optionally:
\itemize{
 \item NORM - The number of reads covering this location in the normal.
 \item NORM_ALT - The number of reads with the alternate base covering this location in the normal.
}
We need to calculate:
\itemize{
 \item tau - A value giving the ratio of weights for selecting a read with the alternate and reference allele.  That is, the bias correction.
 \item CorrectedPhase - A series of factors (typically '0' and '1') indicating which SNPs come from the same allele.
 \item phaseOfMajorAllele - Used to determine which phase is the major allele (i.e., the allele with the highest CN).
}
}

