% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\title{Plot coverage and BAF data in the target region specified.}
\usage{
plot_region(ltgts, sampleBAM, bafs, sampleGC_Bias = NULL,
  sampleCorrection = 1, normalBAM = NULL, normalGC_Bias = NULL,
  normalCorrection = 1, referenceStrandBias = 1, bin_size = 100,
  nbins = 100, pos_mode = c("Concatenate", "Absolute"),
  coverageBiasCorrect = "GPS", BAF_BiasCorrect = "GR", logCov = FALSE,
  to_plot = c("Tumour/Normal", "BAF"), coverage_yLims = ifelse(c(logCov,
  logCov), c(-1, 1), c(0, 4)))
}
\arguments{
\item{ltgts}{The set of genomic regions to plot.}

\item{sampleBAM}{The location of the BAM file for the Tumour sample.}

\item{bafs}{The BAF GRanges object.  Should only contain information for the tumour sample of interest, otherwise you'll plot the BAFs from all samples in the provided GRange.  So don't do that.}

\item{sampleGC_Bias}{The GCbias object for the Tumour.}

\item{sampleCorrection}{Correction factor for sample effects for the Tumour sample.}

\item{normalBAM}{The location of the BAM file for the Normal sample.}

\item{normalGC_Bias}{The GCbias object for the Normal.}

\item{normalCorrection}{Correction factor for sample effects for the Normal sample.}

\item{referenceStrandBias}{Factor to correct for reference strand bias.  The ratio of weights favouring the alternate strand over the reference strand.}

\item{bin_size}{How large should the bins it's split into be? If NA, coverage is calculated rather than counts in bins.}

\item{nbins}{Instead of specifying a bin size, you can give the desired numbers of bins and have the bin_size set dynamically.  If bin_size is also set to something other than NA, it specifies a floor below which the bin_size is not allowed to fall.  If set to NULL, this argument is ignored.}

\item{pos_mode}{What is the x-axis given to the data.  Absolute simply plots the absolute genomic locations of the data, concatenate lines all bins one after another, ordered by the target region's start positions.}

\item{coverageBiasCorrect}{What types of bias correction to apply to the coverage data.  A string where each letter indicates the type of correction to apply (so empty string = correct for nothing).  Options are G - GC, P - Position, S - sample.}

\item{BAF_BiasCorrect}{What types of bias correction to apply to the BAF data.  A string where each letter indicates the type of correction to apply.  Options are G- GC, R - Reference bias.}

\item{logCov}{Should the coverage data be plotted on a log2 scale?}

\item{to_plot}{Which data to plot and in what order.  Options are 'Tumour/Normal', 'Tumour','Normal' and, 'BAF'.}

\item{coverage_yLims}{The limits on the coverage y-axis.}
}
\value{
Returns a list with the gg plot object containing the desired plot and the data.frame containing the data to be fed into it.
}
\description{
Plots coverage in bin_size bins, after correcting for GC, composition and position effects.
}

