% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_counts_from_GC.R
\name{predict_counts_from_GC}
\alias{predict_counts_from_GC}
\title{Predict number of fragments with 5' end in ranges from GC content}
\usage{
predict_counts_from_GC(tgts, gcBias, summarise = c("Region", "GRange", "BP"))
}
\arguments{
\item{tgts}{A GRanges object giving the locations at which we should predict counts.}

\item{gcBias}{The gcBias object for this sample.}

\item{summarise}{Should we return the number of predicted counts per base pair ('BP'), per entry in tgts ('GRange') or in total ('Region').}
}
\value{
If summarise=='Region' then a list giving the predicted number of forward and reverse counts.  Otherwise returns the tgts object with columns forw and revr giving the predicted counts for that particular range (if summarise is 'GRange') or at each base pair in order (if summarise is 'BP').
}
\description{
Uses the GC bias correction given to calculate how many fragments to expect based on the GC content in windows anchored at locations in the target regions specified.
}

