% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prefix_chr}
\alias{prefix_chr}
\title{Add a prefix to chromosomes}
\usage{
prefix_chr(chrs, prefix = "chr")
}
\arguments{
\item{chrs}{The chromosome names to prefix.}

\item{prefix}{The prefix to prepend.}
}
\value{
The processed chromosome names.
}
\description{
Strips any existing prefix and appends a prefix of your choice.
}

