% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitiseRanges.R
\name{sanitiseRanges}
\alias{sanitiseRanges}
\title{Truncates or drops regions that are within buf of the start/end of chromosomes}
\usage{
sanitiseRanges(regions, genome, buf)
}
\arguments{
\item{regions}{The GRanges object to be sanitised.}

\item{genome}{A BSgenome object from which sequence lengths can be extracted to test for unacceptable coordinates.}

\item{buf}{An integer specifying how close to the start or end of a chromosome coordinates are allowed to be.}
}
\value{
The GRanges object given in regions with ranges trimmed or removed as needed to satisfy the requirements of the function.
}
\description{
Removes those parts of regions that either directly refer to invalid genomic coordinates or are within buf of the start or end of a chromosome.  Mostly useful for excluding locations with bad windows for GC prediction.
}

