% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varString}
\alias{varString}
\title{Replace variables in string}
\usage{
varString(string, ...)
}
\arguments{
\item{string}{The string to replace variables in.}

\item{...}{What to replace with what.}
}
\value{
The parsed string.
}
\description{
An extended version of sprintf that allows named variables to be replaced within a string.
}
\examples{
varString("It's \%day\% today in the month of \%month\%",day="Monday",month="Smarch")
}

